1) updated tariff domain to fixed-point math 2) add new fields to tariff

+++ src/main/java/de/rwth/idsg/bikeman/service/TariffPriceCalculation.java
@@ -2,6 +2,8 @@ package de.rwth.idsg.bikeman.service;
 
 import de.rwth.idsg.bikeman.domain.Transaction;
 
+import java.math.BigDecimal;
+
 /**
  * Created by swam on 20/01/15.
  */
@@ -9,5 +11,5 @@ public interface TariffPriceCalculation {
     
     
     // calculates price in euro
-     double calculate(Transaction transaction);
+     BigDecimal calculate(Transaction transaction);
 }
